+++ CHANGELOG.md
@@ -12,6 +12,7 @@ All notable changes to `homebridge-config-ui-x` will be documented in this file.
 - prevent url-based settings from being changed from a pwa
 - add search bar on settings page for filtering options
 - implement html page titles on section pages (#2579)
+- fix for custom uis showing schema form
 
 ### Other Changes